backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
httpx==0.27.0
anthropic==0.34.0
onnxruntime==1.19.2
faiss-cpu==1.8.0
Pillow==10.4.0
numpy==1.26.4
aiofiles==24.1.0
